third_party/WebKit/LayoutTests/svg/dom/SVGAnimatedInteger-expected.txt

This test checks the SVGAnimatedInteger API - utilizing the targetX property of SVGFEConvolveMatrix

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check initial targetX value
PASS feConvolveMatrix.targetX.toString() is "[object SVGAnimatedInteger]"
PASS typeof(feConvolveMatrix.targetX.baseVal) is "number"
PASS feConvolveMatrix.targetX.baseVal is 0

Check that integers are static, caching value in a local variable and modifying it, should have no effect
PASS numRef is 100
PASS feConvolveMatrix.targetX.baseVal is 0

Check assigning various valid and invalid values
PASS feConvolveMatrix.targetX.baseVal = -1 is -1
PASS feConvolveMatrix.targetX.baseVal = 300 is 300
PASS feConvolveMatrix.targetX.baseVal = 'aString' is 'aString'
PASS feConvolveMatrix.targetX.baseVal is 0
PASS feConvolveMatrix.targetX.baseVal = feConvolveMatrix is feConvolveMatrix
PASS feConvolveMatrix.targetX.baseVal is 0
PASS feConvolveMatrix.targetX.baseVal = 300 is 300

Check that the targetX value remained 300
PASS feConvolveMatrix.targetX.baseVal is 300
PASS successfullyParsed is true

TEST COMPLETE
